% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_event.R
\name{add_event_observation}
\alias{add_event_observation}
\title{Append event observations}
\usage{
add_event_observation(
  nif,
  sdtm,
  domain,
  testcd = NULL,
  event_filter,
  event_diff = FALSE,
  analyte = NULL,
  parent = NULL,
  metabolite = FALSE,
  cmt = NULL,
  subject_filter = "!ACTARMCD \%in\% c('SCRdeeNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  DTC_field = NULL,
  keep = NULL,
  debug = FALSE,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{sdtm}{A sdtm object.}

\item{domain}{The domain code as character.}

\item{testcd}{The testcode from xxTESTCD where xx is the domain code, as
character. Not used if NULL.}

\item{event_filter}{A filter term to characterize the events to extract, as
character.}

\item{event_diff}{Only retain events where there is a change from the event
filter evaluating to TRUE after being FALSE in the previous observation,
i.e., apply time differentiation to the observation. As logical. Defaults
to FALSE.}

\item{analyte}{The name for the analyte. Defaults to the 'EV_testcd', if
NULL.}

\item{parent}{The name of the parent analyte for the observation as
character. Defaults to the value of 'analyte' if NULL.}

\item{metabolite}{observation is a metabolite, as logical.}

\item{cmt}{The compartment for the observation, as numeric.}

\item{subject_filter}{The filter to be applied to the DM domain, as
character.}

\item{observation_filter}{The filtering to apply to the observation source
data, as character.}

\item{DTC_field}{The field to use as the date-time code for the observation.
Defaults to 'xxDTC', with xx the domain name, if NULL.}

\item{keep}{Columns to keep, as character.}

\item{debug}{Add debug information, as logical.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}
}
\value{
A nif object.
}
\description{
`r lifecycle::badge("experimental")`
}
