% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_plot_surv.R
\name{kmplot}
\alias{kmplot}
\title{Kaplan-Meier plot for events}
\usage{
kmplot(
  nif,
  analyte,
  dose = NULL,
  group = NULL,
  convert_tafd_h_to_d = TRUE,
  title = NULL,
  y_label = NULL,
  show_censoring = TRUE,
  show_risk_table = TRUE,
  show_ci = TRUE,
  silent = NULL
)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte.}

\item{dose}{The dose, defaults to all doses, if NULL.}

\item{group}{Grouping variable.}

\item{convert_tafd_h_to_d}{Convert the TAFD field from hours to days,
defaults to TRUE.}

\item{title}{The plot title, defaults to none if NULL.}

\item{y_label}{The y axis label, defaults to the analyte, if NULL.}

\item{show_censoring}{Add censor mark, as logical.}

\item{show_risk_table}{Add risk table, as logical.}

\item{show_ci}{Add confidence interval, as logical.}

\item{silent}{Suppress messages, as logical. Defaults to nif_option setting
if NULL.}
}
\value{
A ggsurvfit (ggplot) object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
