% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifevent-package.R
\docType{package}
\name{nifevent-package}
\alias{nifevent}
\alias{nifevent-package}
\title{nifevent: Event Observations for NIF Data Sets}
\description{
Add and explore event observations.
}
\author{
\strong{Maintainer}: Rainer Strotmann \email{rainer.strotmann@mailbox.org} (\href{https://orcid.org/0000-0003-0543-7984}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rainer Strotmann \email{rainer.strotmann@mailbox.org} (\href{https://orcid.org/0000-0003-0543-7984}{ORCID}) [copyright holder]
}

}
\keyword{internal}
