% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nif_plot_surv.R
\name{kmplot}
\alias{kmplot}
\title{Kaplan-Meier plot for events}
\usage{
kmplot(nif, analyte, dose = NULL, group = NULL)
}
\arguments{
\item{nif}{A nif object.}

\item{analyte}{The analyte.}

\item{dose}{The dose, defaults to all doses, if NULL.}

\item{group}{Grouping variable.}
}
\value{
A ggplot object.
}
\description{
`r lifecycle::badge("experimental")`
}
